fix: countries graphql queries
